+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Optimize icon type detection (iOS/MD).
 - Optimize logic for merging user props with defaults.
 - Minor refactorization to the main component, more comments.
+- Change folder names "bundle" and "icon" to "bundles" and "icons", for consistency with the ion-icon web-component.
 
 ## \[0.1.4] - 2019-01-31